fix(attention): review polish — honest cost ceiling, vocabulary constants, seam split
The label pass's comment states the true worst case (~200 gated
single-row reads on a feed whose every lane is full) instead of 'a few
dozen', and the batched resolver is filed as the follow-up it names
(#3051). The error posture is documented as deliberate: a database
failure must not render as an absent label, because absence means 'you
may not read this' — the same lie face() refuses. The resolver's switch
uses the datasource record vocabulary rather than fresh literals, and
the face builders moved beside their second caller (attentionnames.go),
which also keeps attentionseam.go under the file cap.
